Some refactoring, cleaning and removal of old code
