+++ app/javascript/components/avatars/initialize/grab_elements.js
@@ -1,15 +1,7 @@
 export function grabElements() {
   return(
     {
-      imgBase: document.getElementById("face"),
-      imgHair: document.getElementById("hair"),
-      imgMouth: document.getElementById("mouth"),
-      imgEyes: document.getElementById("eyes"),
-      imgEyebrows: document.getElementById("eyebrows"),
-      imgNose: document.getElementById("nose"),
-      imgCloth: document.getElementById("cloth") ?? '',
-      imgAcessory: document.getElementById("acessory") ?? '',
-      masterLayer: document.querySelector(".result")
+      masterLayer: document.getElementById("avatarCanvas")
     }
   );
 }
@@ -21,19 +13,4 @@ export async function getAvatar() {
   const response = await fetch(`${URL}/${id}`);
   avatar = await response.json();
   return avatar;
-}
-
-export function setCurrentAssets(avDom) {
-  return(
-    {
-      base: avDom.imgBase.src.slice(avDom.imgBase.src.lastIndexOf("/") + 1),
-      hair: avDom.imgHair.src.slice(avDom.imgHair.src.lastIndexOf("/") + 1),
-      mouth: avDom.imgMouth.src.slice(avDom.imgMouth.src.lastIndexOf("/") + 1),
-      eyes: avDom.imgEyes.src.slice(avDom.imgEyes.src.lastIndexOf("/") + 1),
-      eyebrows: avDom.imgEyebrows.src.slice(avDom.imgEyebrows.src.lastIndexOf("/") + 1),
-      nose: avDom.imgNose.src.slice(avDom.imgNose.src.lastIndexOf("/") + 1),
-      clothes: avDom.imgCloth.src.slice(avDom.imgCloth.src.lastIndexOf("/") + 1),
-      acessory: avDom.imgAcessory.src.slice(avDom.imgAcessory.src.lastIndexOf("/") + 1)
-    }
-  );
 }
