+++ app/javascript/components/avatars/update_canvas.js
@@ -58,43 +58,3 @@ export function initCanvas(dom, avatar) {
     clothe: layers.clothe
   }
 }
-
-
-export function initializeCanvas(dom, assets, index) {
-  const context          = dom.masterLayer.getContext("2d");
-  dom.masterLayer.width  = dom.imgBase.width;
-  dom.masterLayer.height = dom.imgBase.height;
-  const layers = initializeLayers(dom, context, assets, index);
-  context.drawImage(layers.base.info.layer, 0, 0);
-  context.drawImage(layers.eyes.info.layer, 0, 0);
-  context.drawImage(layers.hair.info.layer, 0, 0);
-  context.drawImage(layers.eyebrows.info.layer, 0, 0);
-  context.drawImage(layers.acessory.info.layer, 0, 0);
-  context.drawImage(layers.clothe.info.layer, 0, 0);
-  return {
-    layers,
-    context
-  }
-}
-
-export function updateCanvas(dom, context, layers) {
-  // layers.base.draw();
-  // context.drawImage(layers.base.info.layer, 0, 0);
-  // layers.eyes.draw();
-  // context.drawImage(layers.eyes.info.layer, 0, 0);
-  // layers.hair.draw();
-  // context.drawImage(layers.hair._info.layer, 0, 0);
-  // layers.eyebrows.draw();
-  // context.drawImage(layers.eyebrows.info.layer, 0, 0);
-  // layers.acessory.draw();
-  // context.drawImage(layers.acessory.info.layer, 0, 0);
-  // layers.clothe.draw();
-  // context.drawImage(layers.clothe.info.layer, 0, 0);
-  //context = dom.masterLayer.getContext('2d');
-  context.drawImage(layers.base.info.layer, 0, 0);
-  context.drawImage(layers.eyes.info.layer, 0, 0);
-  context.drawImage(layers.eyebrows.info.layer, 0, 0);
-  context.drawImage(layers.hair.info.layer, 0, 0);
-  context.drawImage(layers.acessory.info.layer, 0, 0);
-  context.drawImage(layers.clothe.info.layer, 0, 0);
-}
