+++ app/views/avatars/index.html.erb
@@ -1,20 +1,6 @@
 
-<div class="avatar-data">
-  <%= content_tag :div, class: "avatar_information", data: { bases: @bases, basecolors: @base_colors, eyes: @eyes, eyecolors: @eye_colors, hairs: @hairs, haircolors: @colorized_hairs, mouths: @mouths, eyebrows: @eyebrows, eyebrowcolors: @eyebrow_colors, noses: @noses, clothes: @clothes, acessories: @acessories, acessorycolors: @acessory_colors, gender: @gender } do %>
-  <% end %>
   <%= content_tag :div, class: "avdata", data: { id: @avatar.id } do %>
   <% end %>
-  <%= image_tag "/avatar/#{@base}", id: 'face', class: 'hidden-content' %>
-  <%= image_tag "/avatar/#{@hair}", id: 'hair', class: 'hidden-content'%>
-  <%= image_tag "/avatar/#{@mouth}", id: 'mouth', class: 'hidden-content'%>
-  <%= image_tag "/avatar/#{@nose}", id: 'nose', class: 'hidden-content' %>
-  <%= image_tag "/avatar/#{@eye}", id: 'eyes', class: 'hidden-content' %>
-  <%= image_tag "/avatar/#{@eyebrow}", id: 'eyebrows', class: 'hidden-content' %>
-  <%= image_tag "/avatar/#{@cloth}", id: 'cloth', class: 'hidden-content' if @cloth %>
-  <%= image_tag '', id: 'cloth', class: 'hidden-content' unless @cloth %>
-  <%= image_tag "/avatar/#{@acessory}", id: 'acessory', class: 'hidden-content' if @acessory %>
-  <%= image_tag '', id: 'acessory', class: 'hidden-content' unless @acessory %>
-</div>
 <div class="general-container">
 <div class="avatar-edit-container">
   <div class="card-edit-left">
@@ -98,7 +84,7 @@
 <!--       <span class="Btn-male absolute"><%= image_tag "male.png" %></span>
       <span class="Btn-female absolute"><%= image_tag "fem.png" %></span> -->
       <div class="canvas-img-box">
-        <canvas class="result"></canvas>
+        <canvas id="avatarCanvas"></canvas>
         <div class="profile-name topos-12">
           <span>
             <%= @user.username %>
@@ -184,20 +170,8 @@
             <span class="pointer"><</span>
           </div>
           Clothes
-<!--           <div class="arrow-left-filler second-card-arrow-pos-left">
-            <span class="pointer"></span>
-          </div> -->
         </div>
       </div>
-<!--       <div class="gender-btn absolute">
-        <%= show_svg('gender.svg') %>
-        <div class="Btn-female gender-txt to-right pointer">
-          <span>Female</span>
-        </div>
-        <div class="Btn-male gender-txt to-left pointer">
-          <span>Male</span>
-        </div>
-      </div> -->
 
     </div>
   </div>
